+++ invoke/runners.py
@@ -1264,7 +1264,8 @@ class Local(Runner):
                 # Use execve for bare-minimum "exec w/ variable # args + env"
                 # behavior. No need for the 'p' (use PATH to find executable)
                 # for now.
-                # TODO: see if subprocess is using equivalent of execvp...
+                # NOTE: stdlib subprocess (actually its posix flavor, which is
+                # written in C) uses either execve or execv, depending.
                 os.execve(shell, [shell, "-c", command], env)
         else:
             self.process = Popen(
@@ -1318,8 +1319,16 @@ class Local(Runner):
             return self.process.returncode
 
     def stop(self):
-        # No explicit close-out required (so far).
-        pass
+        # If we opened a PTY for child communications, make sure to close() it,
+        # otherwise long-running Invoke-using processes exhaust their file
+        # descriptors eventually.
+        if self.using_pty:
+            try:
+                os.close(self.parent_fd)
+            except:
+                # If something weird happened preventing the close, there's
+                # nothing to be done about it now...
+                pass
 
 
 class Result(object):